Typescript builds on JS foundation
