+++ src/languages/typescript.js
@@ -5,56 +5,20 @@ Contributors: Ike Ku <dempfi@yahoo.com>
 Description: TypeScript is a strict superset of JavaScript
 Website: https://www.typescriptlang.org
 Category: common, scripting
-
 */
 
 import * as ECMAScript from "./lib/ecmascript.js";
-import * as regex from "../lib/regex.js";
+import javascript from "./javascript.js";
 
+/** @type LanguageFn */
 export default function(hljs) {
-
-  /**
-   * Takes a string like "<Booger" and checks to see
-   * if we can find a matching "</Booger" later in the
-   * content.
-   * @param {RegExpMatchArray} match
-   * @param {{after:number}} param1
-   */
-  const hasClosingTag = (match , {after}) => {
-    const tag = match[0].replace("<","</");
-    const pos = match.input.indexOf(tag, after)
-    return pos !== -1;
-  };
-
   var IDENT_RE = ECMAScript.IDENT_RE;
-  var FRAGMENT = {
-    begin: '<>',
-    end: '</>'
+  var NAMESPACE = {
+    beginKeywords: 'namespace', end: /\{/, excludeEnd: true
   };
-  var XML_TAG = {
-    begin: /<[A-Za-z0-9\\._:-]+/,
-    end: /\/[A-Za-z0-9\\._:-]+>|\/>/,
-    isTrulyOpeningTag: (match, response) => {
-      const afterMatchIndex = match[0].length + match.index;
-      const nextChar = match.input[afterMatchIndex]
-      // nested type?
-      // HTML should not include another raw `<` inside a tag
-      // But a type might: `<Array<Array<number>>`, etc.
-      if (nextChar === "<") {
-        response.ignoreMatch();
-        return;
-      }
-      // <something>
-      // This is now either a tag or a type.
-      if (nextChar===">") {
-        // if we cannot find a matching closing tag, then we
-        // will ignore it
-        if (!hasClosingTag(match, {after: afterMatchIndex})) {
-          response.ignoreMatch();
-          return;
-        }
-      }
-    }
+  var INTERFACE = {
+    beginKeywords: 'interface', end: /\{/, excludeEnd: true,
+    keywords: 'interface extends'
   };
   var TYPES = [
     "any",
@@ -89,313 +53,24 @@ export default function(hljs) {
     className: 'meta',
     begin: '@' + IDENT_RE,
   };
-  var NUMBER = {
-    className: 'number',
-    variants: [
-      { begin: '\\b(0[bB][01]+)n?' },
-      { begin: '\\b(0[oO][0-7]+)n?' },
-      { begin: hljs.C_NUMBER_RE + 'n?' }
-    ],
-    relevance: 0
-  };
-  var SUBST = {
-    className: 'subst',
-    begin: '\\$\\{', end: '\\}',
-    keywords: KEYWORDS,
-    contains: [] // defined later
-  };
-  var HTML_TEMPLATE = {
-    begin: 'html`', end: '',
-    starts: {
-      end: '`', returnEnd: false,
-      contains: [
-        hljs.BACKSLASH_ESCAPE,
-        SUBST
-      ],
-      subLanguage: 'xml',
-    }
-  };
-  var CSS_TEMPLATE = {
-    begin: 'css`', end: '',
-    starts: {
-      end: '`', returnEnd: false,
-      contains: [
-        hljs.BACKSLASH_ESCAPE,
-        SUBST
-      ],
-      subLanguage: 'css',
-    }
-  };
-  var TEMPLATE_STRING = {
-    className: 'string',
-    begin: '`', end: '`',
-    contains: [
-      hljs.BACKSLASH_ESCAPE,
-      SUBST
-    ]
-  };
-  var JSDOC_COMMENT = hljs.COMMENT(
-    '/\\*\\*',
-    '\\*/',
-    {
-      relevance : 0,
-      contains : [
-        {
-          className : 'doctag',
-          begin : '@[A-Za-z]+',
-          contains : [
-            {
-              className: 'type',
-              begin: '\\{',
-              end: '\\}',
-              relevance: 0
-            },
-            {
-              className: 'variable',
-              begin: IDENT_RE + '(?=\\s*(-)|$)',
-              endsParent: true,
-              relevance: 0
-            },
-            // eat spaces (not newlines) so we can find
-            // types or variables
-            {
-              begin: /(?=[^\n])\s/,
-              relevance: 0
-            },
-          ]
-        }
-      ]
-    }
-  );
-  var COMMENT = {
-    className: "comment",
-    variants: [
-      JSDOC_COMMENT,
-      hljs.C_BLOCK_COMMENT_MODE,
-      hljs.C_LINE_COMMENT_MODE
-    ]
-  };
-  SUBST.contains = [
-    hljs.APOS_STRING_MODE,
-    hljs.QUOTE_STRING_MODE,
-    HTML_TEMPLATE,
-    CSS_TEMPLATE,
-    TEMPLATE_STRING,
-    NUMBER,
-    hljs.REGEXP_MODE
-  ];
-  var SUBST_AND_COMMENTS = SUBST.contains.concat(COMMENT);
-  var PARAMS_CONTAINS = SUBST_AND_COMMENTS.concat([
-    // eat recursive parens in sub expressions
-    { begin: /\(/,
-      end: /\)/,
-      keywords: KEYWORDS,
-      contains: ["self"].concat(SUBST_AND_COMMENTS)
-    },
-  ]);
-  // TODO: remove, this is older
-  var ARGUMENTS =
-  {
-    begin: '\\(',
-    end: /\)/,
-    keywords: KEYWORDS,
-    contains: [
-      'self',
-      hljs.QUOTE_STRING_MODE,
-      hljs.APOS_STRING_MODE,
-      NUMBER
-    ]
-  };
-  var PARAMS = {
-    className: 'params',
-    begin: /\(/, end: /\)/,
-    excludeBegin: true,
-    excludeEnd: true,
-    keywords: KEYWORDS,
-    contains: PARAMS_CONTAINS.concat(DECORATOR)
-    // [
-    //   hljs.C_LINE_COMMENT_MODE,
-    //   hljs.C_BLOCK_COMMENT_MODE,
-    //   DECORATOR,
-    //   ARGUMENTS
-    // ]
-  };
 
-  return {
+  const tsLanguage = javascript(hljs);
+
+  // this should update anywhere keywords is used since
+  // it will be the same actual JS object
+  Object.assign(tsLanguage.keywords, KEYWORDS);
+
+  Object.assign(tsLanguage, {
     name: 'TypeScript',
-    aliases: ['ts'],
-    keywords: KEYWORDS,
-    contains: [
-      hljs.SHEBANG(),
-      {
-        className: 'meta',
-        begin: /^\s*['"]use strict['"]/
-      },
-      hljs.APOS_STRING_MODE,
-      hljs.QUOTE_STRING_MODE,
-      HTML_TEMPLATE,
-      CSS_TEMPLATE,
-      TEMPLATE_STRING,
-      COMMENT,
-      NUMBER,
-      { // object attr container
-        begin: regex.concat(/[{,\n]\s*/,
-          // we need to look ahead to make sure that we actually have an
-          // attribute coming up so we don't steal a comma from a potential
-          // "value" container
-          //
-          // NOTE: this might not work how you think.  We don't actually always
-          // enter this mode and stay.  Instead it might merely match `,
-          // <comments up next>` and then immediately end after the , because it
-          // fails to find any actual attrs. But this still does the job because
-          // it prevents the value contain rule from grabbing this instead and
-          // prevening this rule from firing when we actually DO have keys.
-          regex.lookahead(regex.concat(
-            // we also need to allow for multiple possible comments inbetween
-            // the first key:value pairing
-            /(((\/\/.*)|(\/\*(.|\n)*\*\/))\s*)*/,
-            IDENT_RE + '\\s*:'))),
-        relevance: 0,
-        contains: [
-          {
-            className: 'attr',
-            begin: IDENT_RE + regex.lookahead('\\s*:'),
-            relevance: 0,
-          },
-        ]
-      },
-      { // "value" container
-        begin: '(' + hljs.RE_STARTERS_RE + '|\\b(case|return|throw)\\b)\\s*',
-        keywords: 'return throw case',
-        contains: [
-          COMMENT,
-          hljs.REGEXP_MODE,
-          {
-            className: 'function',
-            // we have to count the parens to make sure we actually have the
-            // correct bounding ( ) before the =>.  There could be any number of
-            // sub-expressions inside also surrounded by parens.
-            begin: '(\\([^(]*' +
-              '(\\([^(]*' +
-                '(\\([^(]*' +
-                '\\))?' +
-              '\\))?' +
-            '\\)|' + hljs.UNDERSCORE_IDENT_RE + ')\\s*=>', returnBegin: true,
-            end: '\\s*=>',
-            contains: [
-              {
-                className: 'params',
-                variants: [
-                  {
-                    begin: hljs.UNDERSCORE_IDENT_RE
-                  },
-                  {
-                    className: null,
-                    begin: /\(\s*\)/,
-                    skip: true
-                  },
-                  {
-                    begin: /\(/, end: /\)/,
-                    excludeBegin: true, excludeEnd: true,
-                    keywords: KEYWORDS,
-                    contains: PARAMS_CONTAINS
-                  }
-                ]
-              }
-            ]
-          },
-          { // could be a comma delimited list of params to a function call
-            begin: /,/, relevance: 0,
-          },
-          {
-            className: '',
-            begin: /\s/,
-            end: /\s*/,
-            skip: true,
-          },
-          { // JSX
-            variants: [
-              { begin: FRAGMENT.begin, end: FRAGMENT.end },
-              {
-                begin: XML_TAG.begin,
-                // we carefully check the opening tag to see if it truly
-                // is a tag and not a false positive
-                'on:begin': XML_TAG.isTrulyOpeningTag,
-                end: XML_TAG.end
-              }
-            ],
-            subLanguage: 'xml',
-            contains: [
-              {
-                begin: XML_TAG.begin, end: XML_TAG.end, skip: true,
-                contains: ['self']
-              }
-            ]
-          }
-        ],
-        relevance: 0
-      },
-      {
-        className: 'function',
-        beginKeywords: 'function', end: /[\{;]/, excludeEnd: true,
-        keywords: KEYWORDS,
-        contains: [
-          'self',
-          hljs.inherit(hljs.TITLE_MODE, { begin: IDENT_RE }),
-          PARAMS
-        ],
-        illegal: /%/,
-        relevance: 0 // () => {} is more typical in TypeScript
-      },
-      // hack: prevents detection of keywords in some circumstances
-      // .keyword()
-      // $keyword = x
-      {
-        variants: [
-          { begin: '\\.' + IDENT_RE },
-          { begin: '\\$' + IDENT_RE }
-        ],
-        relevance: 0
-      },
-      { // ES6 class
-        className: 'class',
-        beginKeywords: 'class', end: /[{;=]/, excludeEnd: true,
-        illegal: /[:"\[\]]/,
-        contains: [
-          {beginKeywords: 'extends'},
-          hljs.UNDERSCORE_TITLE_MODE
-        ]
-      },
-      {
-        beginKeywords: 'constructor', end: /[\{;]/, excludeEnd: true,
-        contains: [
-          'self',
-          PARAMS
-        ]
-      },
-      {
-        begin: '(get|set)\\s+(?=' + IDENT_RE + '\\()',
-        end: /{/,
-        keywords: "get set",
-        contains: [
-          hljs.inherit(hljs.TITLE_MODE, {begin: IDENT_RE}),
-          { begin: /\(\)/ }, // eat to avoid empty params
-          PARAMS
-        ]
+    aliases: ['ts']
+  });
 
-      },
-      {
-        beginKeywords: 'namespace', end: /\{/, excludeEnd: true
-      },
-      {
-        beginKeywords: 'interface', end: /\{/, excludeEnd: true,
-        keywords: 'interface extends'
-      },
-      {
-        begin: /\$[(.]/ // relevance booster for a pattern common to JS libs: `$(something)` and `$.something`
-      },
-      DECORATOR,
-      ARGUMENTS
-    ]
-  };
-}
+  tsLanguage.exports.PARAMS_CONTAINS.push(DECORATOR);
+  tsLanguage.contains = tsLanguage.contains.concat([
+    DECORATOR,
+    NAMESPACE,
+    INTERFACE,
+  ]);
+
+  return tsLanguage;
+  }
